FixUps and Polished: - Added Password Error PopUp - Added Password Mismatch PopUp - Added Username Already Taken PopUp - Added Cannot Delete Self PopUp - Removed Footer - Added Restricted Search to Vendedores - Added Restricted Search to Asesores - Removed unnecesary Labels - Fixed Dropdown empty option bug

+++ HoldingEmpresas/src/main/java/Programacion2/HoldingEmpresas/controller/DeleteController.java
@@ -25,6 +25,12 @@ public class DeleteController {
 
     @PostMapping("/user")
     public String editUser(@RequestParam Long id, @RequestParam String password, Model model) {
+
+        if(userService.getLoggedUser().getId().equals(id)) {
+            PopUpService.showCannotDeleteSelfPopUp(model);
+            return "edit/user";
+        }
+
         if (userService.checkPassword(password)) {
             userService.delete(id);
             PopUpService.showSuccessPopUp(model);
